合并 generator 和 main 文件
